Data Structure: instruction set

+++ src/ast.rs
@@ -6,14 +6,14 @@ pub enum Expr {
     ELet(Vec<LetEq>, Box<Expr>),
     ELetrec(Vec<LetEq>, Box<Expr>),
     ECase(Box<Expr>, Vec<Alter>),
-    EConstr(i64, i64),
+    EConstr(usize, usize),
     ELam(Vec<Name>, Box<Expr>),
 }
 
 pub type Name = String;
 pub type Program = Vec<ScDef>;
 pub type LetEq = (Name, Expr);
-pub type Alter = (i64, Vec<Name>, Expr);
+pub type Alter = (usize, Vec<Name>, Expr);
 
 #[derive(Debug)]
 pub struct ScDef {
